2018 EVALUACION FINANCIERA PARA FONDUCAR 2011 - 2016
ABSTRACT
This research evaluated financially the Fund of Active and Retired Teachers Employees of the University of Cartagena during the periods 2011 - 2016 using financial analysis and financial indicators to the accounting structure of the entity, with the purpose of knowing the financial position of FONDUCAR, the way how it manages the resources to maintain a solvency and / or effective liquidity that allows the stability of the same and the social benefits of its associates and the community in general. The methodology used throughout this work was the descriptive analytical since historical financial information of FONDUCAR was analyzed and that analysis was described comparing it with entities of a similar nature and operation, thus obtaining fundamental data in the financial growth of the entity such as the financial behavior, and the financial advantages for sustainable development over time.
Key words: financial indicators, financial analysis, liquidity, EVA, EBITDA, Credits.