2016 ESTUDIO DEL PROCESO DE CERTIFICACION DE LOS ESTABLECIMIENTOS GASTRONOMICOS CATEGORIZADOS EN LA CIUDAD DE CARTAGENA BAJO LA NORMA NTS USNA 008
Abstract
The current undergraduate thesis explains the study of the gastronomic establishments located in the city of Cartagena de India supported by the Colombian regulation Restaurants Ranking By Forks -NTS USNA-008-. This norm stipulates the categorization by forks in a rank from one to three. In this particular fact; business which main activity is restoration within the tourist zone of the city are shown.
In its content is developed an analysis of the processes carried out in order to obtain the certification in each restaurant according with the kind of service given. Therefore, the study deepens on the execution of the processes and changes that the restaurants make when they decide to apply the norm under study. The research is performed under the viability of the competitiveness and expansion of the tourist markets in the city and the recognition of the establishments stipulated in the records of the sectorial standardization unit.
Keywords: Quality, Norm NTS USNA, Forks, Gastronomic establishments, Categorization.