2016 ANÁLISIS DE LAS RECOMENDACIONES PARA EL TURISMO ACCESIBLE PROPUESTOS POR LA OMT Y SU APLICACIÓN EN LAS PLAZAS PÚBLICAS DEL CENTRO HISTÓRICO DE CARTAGENA DE INDIAS
The current study looks for analyzing the application of multiples recommendations on accessible tourism in public plazas within the old city in Cartagena de Indias. These recommendations were suggested in 2014 by the World Tourism Organization ( UNWTO) on his “UNWTO recommendations on Accessible Information in Tourism” manual.
We will identify if these places have the adequate characteristics, tools and resources that allow the accessibility to the plazas, recognizing all their advantages and disadvantages and how it influences the whole tourism activity in the city.