2017 “ANÁLISIS DE LA IMPLEMENTACIÓN DE LAS NORMAS INTERNACIONALES DE INFORMACIÓN FINANCIERA EN LA EMPRESA BIOFILM S.A”
The implementation of the International Procedure of Financial Information (NIIF) in the company BIOFILM S.A, focuses principally the transition of the countable procedure between the Decree 2649/93 and the International Procedure of Financial Information (NIIF), besides the impacts on the financial structure, in the business model, in the processes and commercial practices, technological changes among others. The full NIIF contains the beginning for the recognition, measurement, presentation and revelation of the information in financial statements of general intention having in consideration the first reports that BIOLFILM S.A presented under NIIF, starting of from uniform policies to achieve the suitable representation of the economic reality of the company.
In consequence before the generation of this report it makes sense to execute a serious, consistent and coherent process of implementation that allows to establish technical, functional and technological actions needed to analyze and to understand the most relevant aspects of the NIIF as from the objective to provide financial information that is useful to make the reports comparable in the process of capture of decisions of performance and results of the entity.
CLUE WORDS: Financial statements, Globalization, IASB, NIIF, Reasonable Value, Value of Market, Clear Realizable Value.