2017 PROPUESTA DE UN SISTEMA DE CONTROL DE ACTIVOS FIJOS E INVENTARIOS EN LA EMPRESA EG INGENIEROS SAS
The present research aims to propose a control of fixed assets and inventories, according to the International Financial Reporting Standards. The company EG INGENIERO S.A.S. Object of this research has as business model: the design, manufacture, maintenance and assembly in the development of metalworking projects, that is to say that the control in this entity is fundamental. In making a diagnosis between the accountant and the physical relationship of the Inventory and Fixed Assets, three situations were found: inventory overruns, inaccuracy and handling of all assets alike, therefore presenting shortcomings in its current control system. The proposal is aimed at the creation of accounting and administrative policies that allow the correct treatment of information, through formats of control of fixed assets and inventories.
KEY WORDS: Control, Cost, Procedural Manual, System, (Monitoring).