2016 Caracterización de la Inserción Laboral de Estudiantes de Administración Turística y Hotelera de la Fundación Universitaria Colombo Internacional en Cartagena de Indias
Abstract
This research provides significant information about the socioeconomic situation in which the students of VIII and IX semester of the program of Administration of Tourism and Hotel Business are, from the Fundación Universitaria Colombo Internacional, during the I and II period of 2016, and the principal socio-cultural reasons why such a situation occurs, especially when it involves to black or Afrodescendant students from 1 and 2 stratum level. This is finally a huge challenge for both the students and the institution which they are formed as professionals.
Therefore, in this investigation it is set out, first, the demographic and socio-cultural aspects of the students in question; second, it is identified the factors which are associated with unemployment of Afrodescendant student population from UNICOLOBO and the variables that enable the employment of these people and, finally, as a third point, it is proposed a model of labor inclusion for black students from stratum 1 and 2, of this institution of higher education in the Touristic and Cultural District of Cartagena de Indias.
The final reflections of this research focus at, precisely, the challenge that supposes facing this reality by black students in Cartagena de Indias, within a society and an increasingly inequitable labor market, which requires not only professionally qualified persons , but also people who can face this working reality every time less stable and lasting.
Keywords: characterization, employment, tourism, Afrodescendants.